# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: NextJourney REST API
  description: "API REST de NextJourney para gestionar usuarios, destinos, viajes,\
    \ imágenes y reseñas."
  version: v1
servers:
- url: https://localhost:8443
  description: Generated server url
...
paths:
  /api/v1/users/profile/{id}:
    get:
      tags:
      - Usuarios
      operationId: profile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - Usuarios
      operationId: editProfile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: currentPassword
        in: query
        required: false
        schema:
          type: string
      - name: newPassword
        in: query
        required: false
        schema:
          type: string
      - name: confirmPassword
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                user:
                  $ref: "#/components/schemas/UserDTO"
              required:
              - user
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/users/profile:
    get:
      tags:
      - Usuarios
      operationId: profile_1
      parameters: []
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - Usuarios
      operationId: editProfile_1
      parameters:
      - name: currentPassword
        in: query
        required: false
        schema:
          type: string
      - name: newPassword
        in: query
        required: false
        schema:
          type: string
      - name: confirmPassword
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                user:
                  $ref: "#/components/schemas/UserDTO"
              required:
              - user
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/images/{id}/media:
    get:
      tags:
      - Imágenes
      operationId: getImageFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - Imágenes
      operationId: replaceImageFile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/users/profile/image:
    post:
      tags:
      - Usuarios
      operationId: uploadProfileImage
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - Usuarios
      operationId: deleteProfileImage
      parameters: []
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/users/profile/{id}/image:
    post:
      tags:
      - Usuarios
      operationId: uploadProfileImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - Usuarios
      operationId: deleteProfileImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/images/{id}:
    get:
      tags:
      - Imágenes
      operationId: getImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
...
  /api/v1/users/{id}:
    delete:
      tags:
      - Usuarios
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        password:
          type: string
          writeOnly: true
        image:
          $ref: "#/components/schemas/ImageDTO"
        roles:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT